New programs
- observe V3 that fits based on the clustering. The program selects the
best fit between a linear fit or a sat expo, and if the linear is the
best, fits with a power function. The 5 extracted parameters are then
the Diff in and out of corral, the radius of confinement, the residency
time, and the anomalous factor alpha outside of the corral.
- fit multiple, where you can fit all requested data at once
- result tracking V2 to transform the txt files from FIJI to mat files
- pull data to group data of the same type
- new version of compare data